# docs/swagger.yaml
basePath: /api/v1
definitions:
  domain.GetAllSubscriptionsResponse:
    properties:
      meta:
        properties:
          limit:
            type: integer
          offset:
            type: integer
          total:
            type: integer
        type: object
      subscriptions:
        items:
          $ref: '#/definitions/domain.Subscription'
        type: array
    type: object
  domain.Subscription:
    properties:
      created_at:
        type: string
      end_date:
        type: string
      id:
        type: string
      price:
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  domain.SubscriptionRequest:
    properties:
      period:
        enum:
        - 1
        - 3
        - 6
        - 12
        type: integer
      price:
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    required:
    - price
    - service_name
    - start_date
    - user_id
    type: object
host: localhost:8080
info:
  contact: {}
  description: Сервис для агрегации данных об онлайн-подписках пользователей
  title: Subscriptions Service
  version: "1.0"
paths:
  /subscriptions:
    get:
      consumes:
      - application/json
      description: Возвращает список подписок
      parameters:
      - description: Количество записей на странице (по умолчанию = 10)
        in: query
        name: limit
        type: integer
      - description: Смещение (по умолчанию = 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Список подписок с мета-данными
          schema:
            items:
              $ref: '#/definitions/domain.GetAllSubscriptionsResponse'
            type: array
        "400":
          description: Некорректные параметры пагинации
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить список подписок
      tags:
      - subscriptions
    post:
      consumes:
      - application/json
      description: Добавляет новую подписку
      parameters:
      - description: Данные подписки
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/domain.SubscriptionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: string
        "400":
          description: Невалидные данные
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Подписка уже существует
          schema:
            type: string
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Добавить подписку
      tags:
      - subscriptions
  /subscriptions/{id}:
    delete:
      consumes:
      - application/json
      description: Удаляет подписку
      parameters:
      - description: ID подписки (UUID)
        in: path
        name: subscription_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "404":
          description: Невалидный ID
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Удаление подписки
      tags:
      - subscriptions
    get:
      consumes:
      - application/json
      description: Возвращает подписку
      parameters:
      - description: ID подписки (UUID)
        in: path
        name: subscription_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Subscription'
        "400":
          description: Невалидный ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Подписка не найдена
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить подписку по ID
      tags:
      - subscriptions
    put:
      consumes:
      - application/json
      description: Обновляет подписку
      parameters:
      - description: ID подписки (UUID)
        in: path
        name: subscription_id
        required: true
        type: string
      - description: Данные подписки
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/domain.SubscriptionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Subscription'
        "400":
          description: Невалидные данные
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Подписка не найдена
          schema:
            type: string
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Обновление подписки
      tags:
      - subscriptions
  /subscriptions/total_cost/{user_id}:
    get:
      consumes:
      - application/json
      description: Возвращает сумму всех подписок пользователя за период и по названию
        сервис
      parameters:
      - description: ID пользователя (UUID)
        in: path
        name: user_id
        required: true
        type: string
      - description: Начала периода (формат MM-YYYY, например, 01-2025)
        in: query
        name: start_date
        required: true
        type: string
      - description: Конец периода (формат MM-YYYY)
        in: query
        name: end_date
        required: true
        type: string
      - description: Фильт по названию сервиса
        in: query
        name: service_name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: total_cost
          schema:
            additionalProperties:
              type: integer
            type: object
        "400":
          description: Невалидные данные
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Пользователь не найден
          schema:
            type: string
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить сумму всех подписок пользователя
      tags:
      - subscriptions
swagger: "2.0"